Remove deprecated API elements in the com.mongodb package
JAVA-3190
