+++ driver-legacy/src/main/com/mongodb/AggregationOptions.java
@@ -34,36 +34,13 @@ import static java.util.concurrent.TimeUnit.MILLISECONDS;
 public class AggregationOptions {
     private final Integer batchSize;
     private final Boolean allowDiskUse;
-    private final OutputMode outputMode;
     private final long maxTimeMS;
     private final Boolean bypassDocumentValidation;
     private final Collation collation;
 
-    /**
-     * Enumeration to define where the results of the aggregation will be output.
-     * @deprecated There is no replacement for this.  Applications can assume that the driver will use a cursor for server versions
-     * that support it (&gt;= 2.6).  The driver will ignore this as of MongoDB 3.6, which does not support inline results for the aggregate
-     * command.
-     */
-    @Deprecated
-    public enum OutputMode {
-        /**
-         * The output of the aggregate operation is returned inline.
-         */
-        INLINE,
-
-        /**
-         * The output of the aggregate operation is returned using a cursor.
-         *
-         * @mongodb.server.release 2.6
-         */
-        CURSOR
-    }
-
     AggregationOptions(final Builder builder) {
         batchSize = builder.batchSize;
         allowDiskUse = builder.allowDiskUse;
-        outputMode = builder.outputMode;
         maxTimeMS = builder.maxTimeMS;
         bypassDocumentValidation = builder.bypassDocumentValidation;
         collation = builder.collation;
@@ -88,20 +65,6 @@ public class AggregationOptions {
      */
     public Integer getBatchSize() {
         return batchSize;
-    }
-
-    /**
-     * The mode of output for this configuration.
-     *
-     * @return whether the output will be inline or via a cursor, which defaults to {@link OutputMode#CURSOR}
-     * @see OutputMode
-     * @deprecated There is no replacement for this.  Applications can assume that the driver will use a cursor for server versions
-     * that support it (&gt;= 2.6).  The driver will ignore this as of MongoDB 3.6, which does not support inline results for the aggregate
-     * command.
-     */
-    @Deprecated
-    public OutputMode getOutputMode() {
-        return outputMode;
     }
 
     /**
@@ -143,7 +106,6 @@ public class AggregationOptions {
         return "AggregationOptions{"
                 + "batchSize=" + batchSize
                 + ", allowDiskUse=" + allowDiskUse
-                + ", outputMode=" + outputMode
                 + ", maxTimeMS=" + maxTimeMS
                 + ", bypassDocumentValidation=" + bypassDocumentValidation
                 + ", collation=" + collation
@@ -169,7 +131,6 @@ public class AggregationOptions {
     public static class Builder {
         private Integer batchSize;
         private Boolean allowDiskUse;
-        private OutputMode outputMode = OutputMode.CURSOR;
         private long maxTimeMS;
         private Boolean bypassDocumentValidation;
         private Collation collation;
@@ -199,22 +160,6 @@ public class AggregationOptions {
          */
         public Builder allowDiskUse(final Boolean allowDiskUse) {
             this.allowDiskUse = allowDiskUse;
-            return this;
-        }
-
-        /**
-         * The mode of output for this configuration.
-         *
-         * @param mode an {@code OutputMode} that defines how to output the results of the aggregation.
-         * @return {@code this} so calls can be chained
-         * @see OutputMode
-         * @deprecated There is no replacement for this.  Applications can assume that the driver will use a cursor for server versions
-         * that support it (&gt;= 2.6).  The driver will ignore this as of MongoDB 3.6, which does not support inline results for the
-         * aggregate command.
-         */
-        @Deprecated
-        public Builder outputMode(final OutputMode mode) {
-            outputMode = mode;
             return this;
         }
