+++ driver-legacy/src/main/com/mongodb/DB.java
@@ -27,7 +27,6 @@ import com.mongodb.connection.BufferProvider;
 import com.mongodb.lang.Nullable;
 import com.mongodb.operation.BatchCursor;
 import com.mongodb.operation.CommandReadOperation;
-import com.mongodb.operation.CommandWriteOperation;
 import com.mongodb.operation.CreateCollectionOperation;
 import com.mongodb.operation.CreateViewOperation;
 import com.mongodb.operation.DropDatabaseOperation;
@@ -35,7 +34,6 @@ import com.mongodb.operation.ListCollectionsOperation;
 import com.mongodb.operation.ReadOperation;
 import org.bson.BsonDocument;
 import org.bson.BsonDocumentWrapper;
-import org.bson.BsonInt32;
 import org.bson.codecs.BsonDocumentCodec;
 import org.bson.codecs.Codec;
 
@@ -48,11 +46,9 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
 import static com.mongodb.DBCollection.createWriteConcernException;
-import static com.mongodb.MongoCredential.createCredential;
 import static com.mongodb.MongoNamespace.checkDatabaseNameValidity;
 import static com.mongodb.ReadPreference.primary;
 import static com.mongodb.assertions.Assertions.notNull;
-import static java.util.Arrays.asList;
 
 /**
  * A thread-safe client view of a logical database in a MongoDB cluster. A DB instance can be achieved from a {@link MongoClient} instance
@@ -64,7 +60,7 @@ import static java.util.Arrays.asList;
  * }
  * </pre>
  *
- * See {@link Mongo#getDB(String)} for further information about the effective deprecation of this class.
+ * See {@link MongoClient#getDB(String)} for further information about the effective deprecation of this class.
  *
  * @mongodb.driver.manual reference/glossary/#term-database Database
  * @see MongoClient
@@ -72,47 +68,22 @@ import static java.util.Arrays.asList;
 @ThreadSafe
 @SuppressWarnings("deprecation")
 public class DB {
-    private final Mongo mongo;
+    private final MongoClient mongo;
     private final String name;
     private final OperationExecutor executor;
     private final ConcurrentHashMap<String, DBCollection> collectionCache;
-    private final Bytes.OptionHolder optionHolder;
     private final Codec<DBObject> commandCodec;
     private volatile ReadPreference readPreference;
     private volatile WriteConcern writeConcern;
     private volatile ReadConcern readConcern;
 
-    DB(final Mongo mongo, final String name, final OperationExecutor executor) {
+    DB(final MongoClient mongo, final String name, final OperationExecutor executor) {
         checkDatabaseNameValidity(name);
         this.mongo = mongo;
         this.name = name;
         this.executor = executor;
         this.collectionCache = new ConcurrentHashMap<String, DBCollection>();
-        this.optionHolder = new Bytes.OptionHolder(mongo.getOptionHolder());
         this.commandCodec = MongoClient.getCommandCodec();
-    }
-
-    /**
-     * Constructs a new instance of the {@code DB}.
-     *
-     * @param mongo the mongo instance
-     * @param name  the database name - must not be empty and cannot contain spaces
-     * @deprecated Prefer {@link MongoClient#getDB(String)}
-     */
-    @Deprecated
-    public DB(final Mongo mongo, final String name) {
-        this(mongo, name, mongo.createOperationExecutor());
-    }
-
-    /**
-     * Gets the Mongo instance
-     *
-     * @return the mongo instance that this database was created from
-     * @deprecated Use {@link #getMongoClient()} instead
-     */
-    @Deprecated
-    public Mongo getMongo() {
-        return mongo;
     }
 
     /**
@@ -123,11 +94,7 @@ public class DB {
      * @since 3.9
      */
     public MongoClient getMongoClient() {
-        if (mongo instanceof MongoClient) {
-            return (MongoClient) mongo;
-        } else {
-            throw new IllegalStateException("This DB was not created from a MongoClient.  Use getMongo instead");
-        }
+        return mongo;
     }
 
     /**
@@ -197,22 +164,6 @@ public class DB {
     }
 
     /**
-     * <p>Gets a collection with a given name. If the collection does not exist, a new collection is created.</p>
-     *
-     * <p>This class is NOT part of the public API.  Be prepared for non-binary compatible changes in minor releases.</p>
-     *
-     * @param name the name of the collection
-     * @return the collection
-     * @throws IllegalArgumentException if the name is invalid
-     * @see MongoNamespace#checkCollectionNameValidity(String)
-     * @deprecated there is no replacement for this method
-     */
-    @Deprecated
-    protected DBCollection doGetCollection(final String name) {
-        return getCollection(name);
-    }
-
-    /**
      * Gets a collection with a given name.
      *
      * @param name the name of the collection to return
@@ -249,18 +200,6 @@ public class DB {
         } catch (MongoWriteConcernException e) {
             throw createWriteConcernException(e);
         }
-    }
-
-    /**
-     * Returns a collection matching a given string.
-     *
-     * @param collectionName the name of the collection
-     * @return the collection
-     * @deprecated Prefer {@link #getCollection(String)}
-     */
-    @Deprecated
-    public DBCollection getCollectionFromString(final String collectionName) {
-        return getCollection(collectionName);
     }
 
     /**
@@ -579,195 +518,9 @@ public class DB {
         return false;
     }
 
-    /**
-     * Evaluates JavaScript functions on the database server. This is useful if you need to touch a lot of data lightly, in which case
-     * network transfer could be a bottleneck.
-     *
-     * @param code {@code String} representation of JavaScript function
-     * @param args arguments to pass to the JavaScript function
-     * @return result of the command execution
-     * @throws MongoException if the operation failed
-     * @mongodb.driver.manual reference/method/db.eval/ db.eval()
-     * @deprecated The eval command was deprecated in MongoDB 3.0
-     */
-    @Deprecated
-    public CommandResult doEval(final String code, final Object... args) {
-        DBObject commandDocument = new BasicDBObject("$eval", code).append("args", asList(args));
-        return executeCommand(wrap(commandDocument));
-    }
-
-    /**
-     * Calls {@link DB#doEval(java.lang.String, java.lang.Object[]) }. If the command is successful, the "retval" field is extracted and
-     * returned. Otherwise an exception is thrown.
-     *
-     * @param code {@code String} representation of JavaScript function
-     * @param args arguments to pass to the JavaScript function
-     * @return result of the execution
-     * @throws MongoException if the operation failed
-     * @mongodb.driver.manual reference/method/db.eval/ db.eval()
-     * @deprecated The eval command was deprecated in MongoDB 3.0
-     */
-    @Deprecated
-    public Object eval(final String code, final Object... args) {
-        CommandResult result = doEval(code, args);
-        result.throwOnError();
-        return result.get("retval");
-    }
-
-    /**
-     * Helper method for calling a 'dbStats' command. It returns storage statistics for a given database.
-     *
-     * @return result of the execution
-     * @throws MongoException if the operation failed
-     * @mongodb.driver.manual reference/command/dbStats/ Database Stats
-     * @deprecated Prefer {@link #command(DBObject)} to run the dbStats command
-     */
-    @Deprecated
-    public CommandResult getStats() {
-        BsonDocument commandDocument = new BsonDocument("dbStats", new BsonInt32(1)).append("scale", new BsonInt32(1));
-        return executeCommand(commandDocument);
-    }
-
-    /**
-     * Adds or updates a user for this database
-     *
-     * <p>Never create or modify users over an insecure network without the use of TLS.</p>
-     * @param userName the user name
-     * @param password the password
-     * @return the result of executing this operation
-     * @throws MongoException if the operation failed
-     * @mongodb.driver.manual administration/security-access-control/  Access Control
-     * @mongodb.driver.manual reference/command/createUser createUser
-     * @mongodb.driver.manual reference/command/updateUser updateUser
-     * @deprecated Use {@code DB.command} to call either the createUser or updateUser command
-     */
-    @Deprecated
-    public WriteResult addUser(final String userName, final char[] password) {
-        return addUser(userName, password, false);
-    }
-
-    /**
-     * Adds or updates a user for this database
-     *
-     * <p>Never create or modify users over an insecure network without the use of TLS.</p>
-     * @param userName the user name
-     * @param password the password
-     * @param readOnly if true, user will only be able to read
-     * @return the result of executing this operation
-     * @throws MongoException if the operation failed
-     * @mongodb.driver.manual administration/security-access-control/  Access Control
-     * @mongodb.driver.manual reference/command/createUser createUser
-     * @mongodb.driver.manual reference/command/updateUser updateUser
-     * @deprecated Use {@code DB.command} to call either the createUser or updateUser command
-     */
-    @Deprecated
-    public WriteResult addUser(final String userName, final char[] password, final boolean readOnly) {
-        MongoCredential credential = createCredential(userName, getName(), password);
-        boolean userExists = false;
-        try {
-            userExists = executor.execute(new com.mongodb.operation.UserExistsOperation(getName(), userName), primary(), getReadConcern());
-        } catch (MongoCommandException e) {
-            if (e.getCode() != 13) {
-                throw e;
-            }
-        }
-        try {
-            if (userExists) {
-                executor.execute(new com.mongodb.operation.UpdateUserOperation(credential, readOnly, getWriteConcern()), getReadConcern());
-                return new WriteResult(1, true, null);
-            } else {
-                executor.execute(new com.mongodb.operation.CreateUserOperation(credential, readOnly, getWriteConcern()), getReadConcern());
-                return new WriteResult(1, false, null);
-            }
-        } catch (MongoWriteConcernException e) {
-            throw createWriteConcernException(e);
-        }
-    }
-
-    /**
-     * Removes the specified user from the database.
-     *
-     * @param userName user to be removed
-     * @return the result of executing this operation
-     * @throws MongoException if the operation failed
-     * @mongodb.driver.manual administration/security-access-control/  Access Control
-     * @deprecated Use {@code DB.command} to call the dropUser command
-     */
-    @Deprecated
-    public WriteResult removeUser(final String userName) {
-        try {
-            executor.execute(new com.mongodb.operation.DropUserOperation(getName(), userName, getWriteConcern()), getReadConcern());
-            return new WriteResult(1, true, null);
-        } catch (MongoWriteConcernException e) {
-            throw createWriteConcernException(e);
-        }
-    }
-
-    /**
-     * Makes it possible to execute "read" queries on a slave node
-     *
-     * @see ReadPreference#secondaryPreferred()
-     * @deprecated Replaced with {@code ReadPreference.secondaryPreferred()}
-     */
-    @Deprecated
-    public void slaveOk() {
-        addOption(Bytes.QUERYOPTION_SLAVEOK);
-    }
-
-    /**
-     * Adds the given flag to the default query options.
-     *
-     * @param option value to be added
-     * @mongodb.driver.manual ../meta-driver/latest/legacy/mongodb-wire-protocol/#op-query Query Flags
-     * @deprecated Replaced with {@link DBCursor#addOption(int)}
-     */
-    @Deprecated
-    public void addOption(final int option) {
-        optionHolder.add(option);
-    }
-
-    /**
-     * Sets the query options, overwriting previous value.
-     *
-     * @param options bit vector of query options
-     * @mongodb.driver.manual ../meta-driver/latest/legacy/mongodb-wire-protocol/#op-query Query Flags
-     * @deprecated Replaced with {@link DBCursor#setOptions(int)}
-     */
-    @Deprecated
-    public void setOptions(final int options) {
-        optionHolder.set(options);
-    }
-
-    /**
-     * Resets the query options.
-     * @mongodb.driver.manual ../meta-driver/latest/legacy/mongodb-wire-protocol/#op-query Query Flags
-     * @deprecated Replaced with {@link DBCursor#resetOptions()}
-     */
-    @Deprecated
-    public void resetOptions() {
-        optionHolder.reset();
-    }
-
-    /**
-     * Gets the query options
-     *
-     * @return bit vector of query options
-     * @mongodb.driver.manual ../meta-driver/latest/legacy/mongodb-wire-protocol/#op-query Query Flags
-     * @deprecated Replaced with {@link DBCursor#getOptions()}
-     */
-    @Deprecated
-    public int getOptions() {
-        return optionHolder.get();
-    }
-
     @Override
     public String toString() {
         return "DB{name='" + name + '\'' + '}';
-    }
-
-    CommandResult executeCommand(final BsonDocument commandDocument) {
-        return new CommandResult(executor.execute(new CommandWriteOperation<BsonDocument>(getName(), commandDocument,
-                                                                                          new BsonDocumentCodec()), getReadConcern()));
     }
 
     CommandResult executeCommand(final BsonDocument commandDocument, final ReadPreference readPreference) {
@@ -780,12 +533,8 @@ public class DB {
         return executor;
     }
 
-    Bytes.OptionHolder getOptionHolder() {
-        return optionHolder;
-    }
-
     BufferProvider getBufferPool() {
-        return getMongo().getBufferProvider();
+        return getMongoClient().getBufferProvider();
     }
 
     private BsonDocument wrap(final DBObject document) {
