+++ driver-legacy/src/main/com/mongodb/DBCollection.java
@@ -52,7 +52,6 @@ import com.mongodb.operation.MapReduceStatistics;
 import com.mongodb.operation.MapReduceToCollectionOperation;
 import com.mongodb.operation.MapReduceWithInlineResultsOperation;
 import com.mongodb.operation.MixedBulkWriteOperation;
-import com.mongodb.operation.ParallelCollectionScanOperation;
 import com.mongodb.operation.ReadOperation;
 import com.mongodb.operation.RenameCollectionOperation;
 import com.mongodb.operation.UpdateOperation;
@@ -117,7 +116,7 @@ import static java.util.concurrent.TimeUnit.MILLISECONDS;
  * System.out.println(myDoc); }
  * </pre>
  *
- * See {@link Mongo#getDB(String)} for further information about the effective deprecation of this class.
+ * See {@link MongoClient#getDB(String)} for further information about the effective deprecation of this class.
  *
  * @mongodb.driver.manual reference/glossary/#term-collection Collection
  */
@@ -128,12 +127,10 @@ public class DBCollection {
     private final String name;
     private final DB database;
     private final OperationExecutor executor;
-    private final Bytes.OptionHolder optionHolder;
     private final boolean retryWrites;
     private volatile ReadPreference readPreference;
     private volatile WriteConcern writeConcern;
     private volatile ReadConcern readConcern;
-    private List<DBObject> hintFields;
     private DBEncoderFactory encoderFactory;
     private DBDecoderFactory decoderFactory;
     private DBCollectionObjectFactory objectFactory;
@@ -150,10 +147,9 @@ public class DBCollection {
         this.name = name;
         this.database = database;
         this.executor = executor;
-        this.optionHolder = new Bytes.OptionHolder(database.getOptionHolder());
         this.objectFactory = new DBCollectionObjectFactory();
         this.objectCodec = new CompoundDBObjectCodec(getDefaultDBObjectCodec());
-        this.retryWrites = database.getMongo().getMongoClientOptions().getRetryWrites();
+        this.retryWrites = database.getMongoClient().getMongoClientOptions().getRetryWrites();
     }
 
     /**
@@ -164,14 +160,6 @@ public class DBCollection {
      */
     protected DBCollection(final DB database, final String name) {
         this(name, database, database.getExecutor());
-    }
-
-    private static BasicDBList toDBList(final BatchCursor<DBObject> source) {
-        BasicDBList dbList = new BasicDBList();
-        while (source.hasNext()) {
-            dbList.addAll(source.next());
-        }
-        return dbList;
     }
 
     /**
@@ -640,45 +628,6 @@ public class DBCollection {
     /**
      * Select documents in collection and get a cursor to the selected documents.
      *
-     * @param query      the selection criteria using query operators. Omit the query parameter or pass an empty document to return all
-     *                   documents in the collection.
-     * @param projection specifies which fields MongoDB will return from the documents in the result set.
-     * @param numToSkip  number of documents to skip
-     * @param batchSize  see {@link DBCursor#batchSize(int)} for more information
-     * @param options    query options to be used
-     * @return A cursor to the documents that match the query criteria
-     * @mongodb.driver.manual tutorial/query-documents/ Querying
-     * @deprecated use {@link com.mongodb.DBCursor#skip(int)}, {@link com.mongodb.DBCursor#batchSize(int)} and {@link
-     * com.mongodb.DBCursor#setOptions(int)} on the {@code DBCursor} returned from {@link com.mongodb.DBCollection#find(DBObject,
-     * DBObject)}
-     */
-    @Deprecated
-    public DBCursor find(final DBObject query, final DBObject projection, final int numToSkip, final int batchSize,
-                         final int options) {
-        return new DBCursor(this, query, projection, getReadPreference()).batchSize(batchSize).skip(numToSkip).setOptions(options);
-    }
-
-    /**
-     * Select documents in collection and get a cursor to the selected documents.
-     *
-     * @param query      the selection criteria using query operators. Omit the query parameter or pass an empty document to return all
-     *                   documents in the collection.
-     * @param projection specifies which fields MongoDB will return from the documents in the result set.
-     * @param numToSkip  number of documents to skip
-     * @param batchSize  see {@link DBCursor#batchSize(int)} for more information
-     * @return A cursor to the documents that match the query criteria
-     * @mongodb.driver.manual tutorial/query-documents/ Querying
-     * @deprecated use {@link com.mongodb.DBCursor#skip(int)} and {@link com.mongodb.DBCursor#batchSize(int)} on the {@code DBCursor}
-     * returned from {@link com.mongodb.DBCollection#find(DBObject, DBObject)}
-     */
-    @Deprecated
-    public DBCursor find(final DBObject query, final DBObject projection, final int numToSkip, final int batchSize) {
-        return new DBCursor(this, query, projection, getReadPreference()).batchSize(batchSize).skip(numToSkip);
-    }
-
-    /**
-     * Select documents in collection and get a cursor to the selected documents.
-     *
      * @param query the selection criteria using query operators. Omit the query parameter or pass an empty document to return all documents
      *              in the collection.
      * @return A cursor to the documents that match the query criteria
@@ -1020,11 +969,6 @@ public class DBCollection {
         DBObject hint = options.getHint();
         if (hint != null) {
             operation.hint(wrap(hint));
-        } else {
-            String hintString = options.getHintString();
-            if (hintString != null) {
-                operation.hint(new BsonString(hintString));
-            }
         }
         ReadPreference optionsReadPreference = options.getReadPreference();
         ReadConcern optionsReadConcern = options.getReadConcern();
@@ -1062,91 +1006,6 @@ public class DBCollection {
         } catch (MongoWriteConcernException e) {
             throw createWriteConcernException(e);
         }
-    }
-
-    /**
-     * Group documents in a collection by the specified key and performs simple aggregation functions such as computing counts and sums.
-     * This is analogous to a {@code SELECT ... GROUP BY} statement in SQL.
-     *
-     * @param key     specifies one or more document fields to group
-     * @param cond    specifies the selection criteria to determine which documents in the collection to process
-     * @param initial initializes the aggregation result document
-     * @param reduce  specifies an $reduce function, that operates on the documents during the grouping operation
-     * @return a document with the grouped records as well as the command meta-data
-     * @mongodb.driver.manual reference/command/group/ Group Command
-     * @deprecated The group command was deprecated in MongoDB 3.4
-     */
-    @Deprecated
-    public DBObject group(final DBObject key, final DBObject cond, final DBObject initial, final String reduce) {
-        return group(key, cond, initial, reduce, null);
-    }
-
-    /**
-     * Group documents in a collection by the specified key and performs simple aggregation functions such as computing counts and sums.
-     * This is analogous to a {@code SELECT ... GROUP BY} statement in SQL.
-     *
-     * @param key      specifies one or more document fields to group
-     * @param cond     specifies the selection criteria to determine which documents in the collection to process
-     * @param initial  initializes the aggregation result document
-     * @param reduce   specifies an $reduce Javascript function, that operates on the documents during the grouping operation
-     * @param finalize specifies a Javascript function that runs each item in the result set before final value will be returned
-     * @return a document with the grouped records as well as the command meta-data
-     * @mongodb.driver.manual reference/command/group/ Group Command
-     * @deprecated The group command was deprecated in MongoDB 3.4
-     */
-    @Deprecated
-    public DBObject group(final DBObject key, final DBObject cond, final DBObject initial, final String reduce,
-                          @Nullable final String finalize) {
-        return group(key, cond, initial, reduce, finalize, getReadPreference());
-    }
-
-    /**
-     * Group documents in a collection by the specified key and performs simple aggregation functions such as computing counts and sums.
-     * This is analogous to a {@code SELECT ... GROUP BY} statement in SQL.
-     *
-     * @param key            specifies one or more document fields to group
-     * @param cond           specifies the selection criteria to determine which documents in the collection to process
-     * @param initial        initializes the aggregation result document
-     * @param reduce         specifies an $reduce Javascript function, that operates on the documents during the grouping operation
-     * @param finalize       specifies a Javascript function that runs each item in the result set before final value will be returned
-     * @param readPreference {@link ReadPreference} to be used for this operation
-     * @return a document with the grouped records as well as the command meta-data
-     * @mongodb.driver.manual reference/command/group/ Group Command
-     * @deprecated The group command was deprecated in MongoDB 3.4
-     */
-    @Deprecated
-    public DBObject group(final DBObject key, final DBObject cond, final DBObject initial, final String reduce,
-                          @Nullable final String finalize, final ReadPreference readPreference) {
-        return group(new GroupCommand(this, key, cond, initial, reduce, finalize), readPreference);
-    }
-
-    /**
-     * Group documents in a collection by the specified key and performs simple aggregation functions such as computing counts and sums.
-     * This is analogous to a {@code SELECT ... GROUP BY} statement in SQL.
-     *
-     * @param cmd the group command
-     * @return a document with the grouped records as well as the command meta-data
-     * @mongodb.driver.manual reference/command/group/ Group Command
-     * @deprecated The group command was deprecated in MongoDB 3.4
-     */
-    @Deprecated
-    public DBObject group(final GroupCommand cmd) {
-        return group(cmd, getReadPreference());
-    }
-
-    /**
-     * Group documents in a collection by the specified key and performs simple aggregation functions such as computing counts and sums.
-     * This is analogous to a {@code SELECT ... GROUP BY} statement in SQL.
-     *
-     * @param cmd            the group command
-     * @param readPreference {@link ReadPreference} to be used for this operation
-     * @return a document with the grouped records as well as the command meta-data
-     * @mongodb.driver.manual reference/command/group/ Group Command
-     * @deprecated The group command was deprecated in MongoDB 3.4
-     */
-    @Deprecated
-    public DBObject group(final GroupCommand cmd, final ReadPreference readPreference) {
-        return toDBList(executor.execute(cmd.toOperation(getNamespace(), getDefaultDBObjectCodec()), readPreference, getReadConcern()));
     }
 
     /**
@@ -1378,65 +1237,6 @@ public class DBCollection {
     /**
      * Method implements aggregation framework.
      *
-     * @param firstOp       requisite first operation to be performed in the aggregation pipeline
-     * @param additionalOps additional operations to be performed in the aggregation pipeline
-     * @return the aggregation operation's result set
-     * @mongodb.driver.manual core/aggregation-pipeline/ Aggregation
-     * @mongodb.server.release 2.2
-     * @deprecated Use {@link #aggregate(List, AggregationOptions)} instead
-     */
-    @Deprecated
-    @SuppressWarnings("unchecked")
-    public AggregationOutput aggregate(final DBObject firstOp, final DBObject... additionalOps) {
-        List<DBObject> pipeline = new ArrayList<DBObject>();
-        pipeline.add(firstOp);
-        Collections.addAll(pipeline, additionalOps);
-        return aggregate(pipeline);
-    }
-
-    /**
-     * Method implements aggregation framework.
-     *
-     * @param pipeline operations to be performed in the aggregation pipeline
-     * @return the aggregation's result set
-     * @mongodb.driver.manual core/aggregation-pipeline/ Aggregation
-     * @mongodb.server.release 2.2
-     * @deprecated Use {@link #aggregate(List, AggregationOptions)} instead
-     */
-    @Deprecated
-    public AggregationOutput aggregate(final List<? extends DBObject> pipeline) {
-        return aggregate(pipeline, getReadPreference());
-    }
-
-    /**
-     * Method implements aggregation framework.
-     *
-     * @param pipeline       operations to be performed in the aggregation pipeline
-     * @param readPreference the read preference specifying where to run the query
-     * @return the aggregation's result set
-     * @mongodb.driver.manual core/aggregation-pipeline/ Aggregation
-     * @mongodb.server.release 2.2
-     * @deprecated Use {@link #aggregate(List, AggregationOptions, ReadPreference)} )} instead
-     */
-    @Deprecated
-    @SuppressWarnings("unchecked")
-    public AggregationOutput aggregate(final List<? extends DBObject> pipeline, final ReadPreference readPreference) {
-        Cursor cursor = aggregate(pipeline, AggregationOptions.builder().build(), readPreference, false);
-
-        if (cursor == null) {
-            return new AggregationOutput(Collections.<DBObject>emptyList());
-        } else {
-            List<DBObject> results = new ArrayList<DBObject>();
-            while (cursor.hasNext()) {
-                results.add(cursor.next());
-            }
-            return new AggregationOutput(results);
-        }
-    }
-
-    /**
-     * Method implements aggregation framework.
-     *
      * @param pipeline operations to be performed in the aggregation pipeline
      * @param options  options to apply to the aggregation
      * @return the aggregation operation's result set
@@ -1459,17 +1259,7 @@ public class DBCollection {
      */
     public Cursor aggregate(final List<? extends DBObject> pipeline, final AggregationOptions options,
                             final ReadPreference readPreference) {
-        Cursor cursor = aggregate(pipeline, options, readPreference, true);
-        if (cursor == null) {
-            throw new MongoInternalException("cursor can not be null in this context");
-        }
-        return cursor;
-    }
-
-    @SuppressWarnings("deprecation")
-    @Nullable
-    private Cursor aggregate(final List<? extends DBObject> pipeline, final AggregationOptions options, final ReadPreference readPreference,
-                             final boolean returnCursorForOutCollection) {
+        Cursor result;
         notNull("options", options);
         List<BsonDocument> stages = preparePipeline(pipeline);
 
@@ -1483,12 +1273,8 @@ public class DBCollection {
                                                        .collation(options.getCollation());
             try {
                 executor.execute(operation, getReadConcern());
-                if (returnCursorForOutCollection) {
-                    return new DBCursor(database.getCollection(outCollection.asString().getValue()), new BasicDBObject(),
-                            new DBCollectionFindOptions().readPreference(primary()).collation(options.getCollation()));
-                } else {
-                    return null;
-                }
+                result = new DBCursor(database.getCollection(outCollection.asString().getValue()), new BasicDBObject(),
+                        new DBCollectionFindOptions().readPreference(primary()).collation(options.getCollation()));
             } catch (MongoWriteConcernException e) {
                 throw createWriteConcernException(e);
             }
@@ -1497,11 +1283,11 @@ public class DBCollection {
                     .maxTime(options.getMaxTime(MILLISECONDS), MILLISECONDS)
                     .allowDiskUse(options.getAllowDiskUse())
                     .batchSize(options.getBatchSize())
-                    .useCursor(options.getOutputMode() == com.mongodb.AggregationOptions.OutputMode.CURSOR)
                     .collation(options.getCollation());
-            BatchCursor<DBObject> cursor = executor.execute(operation, readPreference, getReadConcern());
-            return new MongoCursorAdapter(new MongoBatchCursorAdapter<DBObject>(cursor));
+            BatchCursor<DBObject> cursor1 = executor.execute(operation, readPreference, getReadConcern());
+            result = new MongoCursorAdapter(new MongoBatchCursorAdapter<DBObject>(cursor1));
         }
+        return result;
     }
 
     /**
@@ -1532,37 +1318,6 @@ public class DBCollection {
         }
 
         return stages;
-    }
-
-    /**
-     * <p>Return a list of cursors over the collection that can be used to scan it in parallel.</p>
-     *
-     * <p>Note: As of MongoDB 2.6, this method will work against a mongod, but not a mongos. </p>
-     *
-     * @param options the parallel scan options
-     * @return a list of cursors, whose size may be less than the number requested
-     * @mongodb.driver.manual reference/command/parallelCollectionScan/ Parallel Collection Scan
-     * @mongodb.server.release 2.6
-     * @since 2.12
-     * @deprecated the parallelCollectionScan command will be removed in MongoDB 4.2
-     */
-    @Deprecated
-    public List<Cursor> parallelScan(final ParallelScanOptions options) {
-        List<Cursor> cursors = new ArrayList<Cursor>();
-        ParallelCollectionScanOperation<DBObject> operation = new ParallelCollectionScanOperation<DBObject>(getNamespace(),
-                                                                                                            options.getNumCursors(),
-                                                                                                            objectCodec)
-                                                                  .batchSize(options.getBatchSize());
-        ReadPreference readPreferenceFromOptions = options.getReadPreference();
-        List<BatchCursor<DBObject>> mongoCursors = executor.execute(operation,
-                                                                    readPreferenceFromOptions != null ? readPreferenceFromOptions
-                                                                                                        : getReadPreference(),
-                getReadConcern());
-
-        for (BatchCursor<DBObject> mongoCursor : mongoCursors) {
-            cursors.add(new MongoCursorAdapter(new MongoBatchCursorAdapter<DBObject>(mongoCursor)));
-        }
-        return cursors;
     }
 
     /**
@@ -1670,31 +1425,6 @@ public class DBCollection {
         } catch (MongoWriteConcernException e) {
             throw createWriteConcernException(e);
         }
-    }
-
-    /**
-     * Get hint fields for this collection (used to optimize queries).
-     *
-     * @return a list of {@code DBObject} to be used as hints.
-     * @mongodb.driver.manual reference/operator/meta/hint/ $hint
-     * @deprecated Prefer {@link DBCursor#hint(DBObject)}
-     */
-    @Deprecated
-    @Nullable
-    public List<DBObject> getHintFields() {
-        return hintFields;
-    }
-
-    /**
-     * Override MongoDB's default index selection and query optimization process.
-     *
-     * @param indexes list of indexes to "hint" or force MongoDB to use when performing the query.
-     * @mongodb.driver.manual reference/operator/meta/hint/ $hint
-     * @deprecated Prefer {@link DBCursor#hint(DBObject)}
-     */
-    @Deprecated
-    public void setHintFields(final List<? extends DBObject> indexes) {
-        this.hintFields = new ArrayList<DBObject>(indexes);
     }
 
     /**
@@ -2086,63 +1816,6 @@ public class DBCollection {
     }
 
     /**
-     * Makes this query ok to run on a slave node
-     *
-     * @deprecated Replaced with {@link ReadPreference#secondaryPreferred()}
-     */
-    @Deprecated
-    public void slaveOk() {
-        addOption(Bytes.QUERYOPTION_SLAVEOK);
-    }
-
-    /**
-     * Adds the given flag to the default query options.
-     *
-     * @param option value to be added
-     * @mongodb.driver.manual reference/method/cursor.addOption/#flags Query Flags
-     * @deprecated Replaced with {@link DBCursor#addOption(int)}
-     */
-    @Deprecated
-    public void addOption(final int option) {
-        optionHolder.add(option);
-    }
-
-    /**
-     * Resets the default query options
-     *
-     * @mongodb.driver.manual reference/method/cursor.addOption/#flags Query Flags
-     * @deprecated Replaced with {@link DBCursor#resetOptions()}
-     */
-    @Deprecated
-    public void resetOptions() {
-        optionHolder.reset();
-    }
-
-    /**
-     * Gets the default query options
-     *
-     * @return bit vector of query options
-     * @mongodb.driver.manual reference/method/cursor.addOption/#flags Query Flags
-     * @deprecated Replaced with {@link DBCursor#getOptions()}
-     */
-    @Deprecated
-    public int getOptions() {
-        return optionHolder.get();
-    }
-
-    /**
-     * Sets the default query options, overwriting previous value.
-     *
-     * @param options bit vector of query options
-     * @mongodb.driver.manual reference/method/cursor.addOption/#flags Query Flags
-     * @deprecated Replaced with {@link DBCursor#setOptions(int)}
-     */
-    @Deprecated
-    public void setOptions(final int options) {
-        optionHolder.set(options);
-    }
-
-    /**
      * Drops (deletes) this collection from the database. Use with care.
      *
      * @throws MongoException if the operation failed
@@ -2157,7 +1830,7 @@ public class DBCollection {
     }
 
     /**
-     * Get the decoder factory for this collection.  A null return value means that the default from MongoOptions is being used.
+     * Get the decoder factory for this collection.  A null return value means that the default from MongoClientOptions is being used.
      *
      * @return the factory
      */
@@ -2166,7 +1839,7 @@ public class DBCollection {
     }
 
     /**
-     * Set a custom decoder factory for this collection.  Set to null to use the default from MongoOptions.
+     * Set a custom decoder factory for this collection.  Set to null to use the default from MongoClientOptions.
      *
      * @param factory the factory to set.
      */
@@ -2182,7 +1855,7 @@ public class DBCollection {
     }
 
     /**
-     * Get the encoder factory for this collection.  A null return value means that the default from MongoOptions is being used.
+     * Get the encoder factory for this collection.  A null return value means that the default from MongoClientOptions is being used.
      *
      * @return the factory
      */
@@ -2191,7 +1864,7 @@ public class DBCollection {
     }
 
     /**
-     * Set a custom encoder factory for this collection.  Set to null to use the default from MongoOptions.
+     * Set a custom encoder factory for this collection.  Set to null to use the default from MongoClientOptions.
      *
      * @param factory the factory to set.
      */
@@ -2320,16 +1993,6 @@ public class DBCollection {
      */
     public void setInternalClass(final String path, final Class<? extends DBObject> aClass) {
         setObjectFactory(objectFactory.update(aClass, asList(path.split("\\."))));
-    }
-
-    /**
-     * Gets the internal class for the given path in the document hierarchy
-     *
-     * @param path the path to map the given Class to
-     * @return the class for a given path in the hierarchy
-     */
-    protected Class<? extends DBObject> getInternalClass(final String path) {
-        return objectFactory.getClassForPath(asList(path.split("\\.")));
     }
 
     @Override
