+++ driver-legacy/src/test/functional/com/mongodb/DBCollectionAggregationTest.java
@@ -19,7 +19,6 @@ package com.mongodb;
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -37,7 +36,6 @@ import static java.util.Arrays.asList;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
@@ -48,75 +46,17 @@ import static org.junit.Assume.assumeTrue;
 public class DBCollectionAggregationTest extends DatabaseTestCase {
 
     @Test
-    public void testAggregation() {
-        DBObject foo = new BasicDBObject("name", "foo").append("count", 5);
-        DBObject bar = new BasicDBObject("name", "bar").append("count", 2);
-        DBObject baz = new BasicDBObject("name", "foo").append("count", 7);
-        collection.insert(foo, bar, baz);
-
-        DBObject projection = new BasicDBObject("name", 1).append("count", 1);
-
-        DBObject group = new BasicDBObject().append("_id", "$name")
-            .append("docsPerName", new BasicDBObject("$sum", 1))
-            .append("countPerName", new BasicDBObject("$sum", "$count"));
-
-        AggregationOutput out = collection.aggregate(
-            Arrays.<DBObject>asList(new BasicDBObject("$project", projection), new BasicDBObject("$group", group)));
-
-        Map<String, DBObject> results = new HashMap<String, DBObject>();
-        for (final DBObject result : out.results()) {
-            results.put((String) result.get("_id"), result);
-        }
-
-        DBObject fooResult = results.get("foo");
-        assertNotNull(fooResult);
-        assertEquals(2, fooResult.get("docsPerName"));
-        assertEquals(12, fooResult.get("countPerName"));
-
-        DBObject barResult = results.get("bar");
-        assertNotNull(barResult);
-        assertEquals(1, barResult.get("docsPerName"));
-        assertEquals(2, barResult.get("countPerName"));
-   }
-
-    @Test
     public void testAggregationCursor() {
         List<DBObject> pipeline = prepareData();
 
         verify(pipeline, AggregationOptions.builder()
                                            .batchSize(1)
-                                           .outputMode(AggregationOptions.OutputMode.CURSOR)
                                            .allowDiskUse(true)
                                            .build());
 
         verify(pipeline, AggregationOptions.builder()
                                            .batchSize(1)
-                                           .outputMode(AggregationOptions.OutputMode.INLINE)
-                                           .allowDiskUse(true)
                                            .build());
-
-        verify(pipeline, AggregationOptions.builder()
-                                           .batchSize(1)
-                                           .outputMode(AggregationOptions.OutputMode.CURSOR)
-                                           .build());
-    }
-
-    @Test
-    public void testInlineAndDollarOut() {
-        String aggCollection = "aggCollection";
-        database.getCollection(aggCollection)
-            .drop();
-        assertEquals(0, database.getCollection(aggCollection)
-                                .count());
-        List<DBObject> pipeline = new ArrayList<DBObject>(prepareData());
-        pipeline.add(new BasicDBObject("$out", aggCollection));
-
-        AggregationOutput out = collection.aggregate(pipeline);
-        assertFalse(out.results()
-                       .iterator()
-                       .hasNext());
-        assertEquals(2, database.getCollection(aggCollection)
-            .count());
     }
 
     @Test
@@ -130,20 +70,18 @@ public class DBCollectionAggregationTest extends DatabaseTestCase {
         List<DBObject> pipeline = new ArrayList<DBObject>(prepareData());
         pipeline.add(new BasicDBObject("$out", aggCollection));
         verify(pipeline, AggregationOptions.builder()
-                                           .outputMode(AggregationOptions.OutputMode.CURSOR)
                                            .build());
         assertEquals(2, database.getCollection(aggCollection)
             .count());
     }
 
     @Test
-    public void testDollarOutOnSecondary() throws UnknownHostException, InterruptedException {
+    public void testDollarOutOnSecondary() throws InterruptedException {
         assumeTrue(clusterIsType(REPLICA_SET));
 
         List<DBObject> pipeline = new ArrayList<DBObject>(prepareData());
         pipeline.add(new BasicDBObject("$out", "aggCollection"));
         AggregationOptions options = AggregationOptions.builder()
-            .outputMode(AggregationOptions.OutputMode.CURSOR)
             .build();
         ServerAddress serverAddress = verify(pipeline, options, ReadPreference.secondary(), collection);
         assertEquals(2, database.getCollection("aggCollection")
@@ -168,38 +106,11 @@ public class DBCollectionAggregationTest extends DatabaseTestCase {
     }
 
     @Test
-    public void testOldAggregationWithOut() {
-        collection.drop();
-        List<DBObject> pipeline = new ArrayList<DBObject>(prepareData());
-        pipeline.add(new BasicDBObject("$out", "aggCollection"));
-        AggregationOutput out = collection.aggregate(pipeline);
-        assertFalse(out.results()
-            .iterator()
-            .hasNext());
-        assertEquals(database.getCollection("aggCollection")
-            .count(), 2);
-    }
-
-    @Test
-    public void testOldAggregationWithOutOnSecondary() throws UnknownHostException, InterruptedException {
-        collection.drop();
-        List<DBObject> pipeline = new ArrayList<DBObject>(prepareData());
-        pipeline.add(new BasicDBObject("$out", "aggCollection"));
-        AggregationOutput out = collection.aggregate(pipeline, ReadPreference.secondary());
-        assertFalse(out.results()
-            .iterator()
-            .hasNext());
-        assertEquals(database.getCollection("aggCollection")
-                             .count(), 2);
-    }
-
-    @Test
     public void testExplain() {
         List<DBObject> pipeline = new ArrayList<DBObject>(prepareData());
         pipeline.add(new BasicDBObject("$out", "aggCollection"));
         CommandResult out = collection.explainAggregate(pipeline, AggregationOptions.builder()
             .allowDiskUse(true)
-            .outputMode(AggregationOptions.OutputMode.CURSOR)
             .build());
         assertTrue(out.keySet()
                       .iterator()
@@ -208,7 +119,7 @@ public class DBCollectionAggregationTest extends DatabaseTestCase {
 
     @Test(expected = IllegalArgumentException.class)
     public void testNullOptions() {
-        collection.aggregate(new ArrayList<DBObject>(), (AggregationOptions) null);
+        collection.aggregate(new ArrayList<DBObject>(), null);
     }
 
     @Test
@@ -251,7 +162,7 @@ public class DBCollectionAggregationTest extends DatabaseTestCase {
     private ServerAddress verify(final List<DBObject> pipeline, final AggregationOptions options, final ReadPreference readPreference,
         final DBCollection collection) {
         Cursor cursor = collection.aggregate(pipeline, options, readPreference);
-        ServerAddress serverAddress = null;
+        ServerAddress serverAddress;
         Map<String, DBObject> results;
         try {
             results = new HashMap<String, DBObject>();
