+++ driver-legacy/src/test/functional/com/mongodb/DBCollectionFunctionalSpecification.groovy
@@ -351,178 +351,6 @@ class DBCollectionFunctionalSpecification extends FunctionalSpecification {
         that distinctValues, containsInAnyOrder(null, 'a', 1, ~['b': 'c'], ~['list': [2, 'd', ~['e': 3]]])
     }
 
-    @SuppressWarnings('UnnecessaryObjectReferences')
-    def insertDataForGroupTests() {
-        collection.drop();
-        collection.save(~['x': 'a']);
-        collection.save(~['x': 'a']);
-        collection.save(~['x': 'a']);
-        collection.save(~['x': 'b']);
-        collection.save(~['x': 'b']);
-        collection.save(~['x': 'b']);
-        collection.save(~['x': 'b']);
-        collection.save(~['x': 'c']);
-    }
-
-    @IgnoreIf({ serverVersionAtLeast(4, 1) })
-    def 'should group by the x field and apply the provided count function'() {
-        given:
-        insertDataForGroupTests()
-
-        when:
-        DBObject result = collection.group(~['x': 1],
-                                           null,
-                                           ~['count': 0],
-                                           'function(o , p){ p.count++; }');
-
-        then:
-        result.size() == 3
-        that result, contains(~['x': 'a', 'count': 3.0D], ~['x': 'b', 'count': 4.0D], ~['x': 'c', 'count': 1.0D]);
-    }
-
-    @IgnoreIf({ serverVersionAtLeast(4, 1) })
-    def 'should group by the x field and apply the provided count function using a GroupCommand'() {
-        given:
-        insertDataForGroupTests()
-        GroupCommand command = new GroupCommand(collection,
-                                                ~['x': 1],
-                                                null,
-                                                ~['count': 0],
-                                                'function(o , p){ p.count++; }',
-                                                null);
-
-        when:
-        DBObject result = collection.group(command);
-
-        then:
-        result.size() == 3
-        that result, contains(~['x': 'a', 'count': 3.0D], ~['x': 'b', 'count': 4.0D], ~['x': 'c', 'count': 1.0D]);
-    }
-
-    @IgnoreIf({ serverVersionAtLeast(4, 1) })
-    def 'should group using a key function in a GroupCommand'() {
-        given:
-        insertDataForGroupTests()
-        GroupCommand command = new GroupCommand(collection,
-                'function(doc){ return {x: doc.x} }',
-                null,
-                ~['count': 0],
-                'function(o , p){ p.count++; }',
-                null);
-
-        when:
-        DBObject result = collection.group(command);
-
-        then:
-        result.size() == 3
-        that result, contains(~['x': 'a', 'count': 3.0D], ~['x': 'b', 'count': 4.0D], ~['x': 'c', 'count': 1.0D]);
-    }
-
-    @IgnoreIf({ serverVersionAtLeast(4, 1) })
-    def 'should group and count only those documents that fulfull the given condition'() {
-        given:
-        insertDataForGroupTests()
-
-        when:
-        def result = collection.group(~['x': 1],
-                                      ~['x': 'b'],
-                                      ~['count': 0],
-                                      'function(o , p){ p.count++; }')
-
-        then:
-        result.size() == 1;
-        that result, contains(~['x': 'b', 'count': 4.0D]);
-    }
-
-    @IgnoreIf({ serverVersionAtLeast(4, 1) })
-    def 'should be able to set a value in the results'() {
-        given:
-        insertDataForGroupTests()
-
-        when:
-        def result = collection.group(~['x': 1],
-                                      ~['x': 'c'],
-                                      ~['valueFound': false],
-                                      'function(o, p){ p.valueFound = true; }');
-
-        then:
-        result.size() == 1;
-        that result, contains(~['x': 'c', 'valueFound': true]);
-    }
-
-    @IgnoreIf({ serverVersionAtLeast(4, 1) })
-    def 'should return results set by the function, based on condition supplied without key'() {
-        given:
-        insertDataForGroupTests()
-
-        when:
-        def result = collection.group(null,
-                                      ~['x': 'b'],
-                                      ~['valueFound': 0],
-                                      'function(o, p){ p.valueFound = true; }');
-
-        then:
-        result.size() == 1
-        that result, contains(~['valueFound': true]);
-    }
-
-    @IgnoreIf({ serverVersionAtLeast(4, 1) })
-    def 'finalize function should replace the whole result document if a document is the return type of finalize function'() {
-        given:
-        insertDataForGroupTests()
-
-        when:
-        def result = collection.group(~['x': 1],
-                                      ~['x': 'b'],
-                                      ~['valueFound': 0],
-                                      'function(o, p){ p.valueFound = true; }',
-                                      'function(o){ return { a:1 } }');
-
-        then:
-        result.size() == 1;
-        that result, contains(~['a': 1.0D]);
-    }
-
-    @IgnoreIf({ serverVersionAtLeast(4, 1) })
-    def 'finalize function should modify the result document'() {
-        given:
-        insertDataForGroupTests()
-
-        when:
-        def result = collection.group(~['x': 1],
-                                      ~['x': 'b'],
-                                      ~['valueFound': 0],
-                                      'function(o, p){ p.valueFound = true; }',
-                                      'function(o){ o.a = 1 }');
-
-        then:
-        result.size() == 1;
-        that result, contains(~['x': 'b', 'valueFound': true, 'a': 1.0D]);
-    }
-
-    @IgnoreIf({ serverVersionAtLeast(4, 1) })
-    def 'should throw an Exception if no initial document provided'() {
-        when:
-        collection.group(~['x': 1],
-                         ~['x': 'b'],
-                         null,
-                         'function(o, p){ p.valueFound = true; }');
-
-        then:
-        thrown(IllegalArgumentException)
-    }
-
-    @IgnoreIf({ serverVersionAtLeast(4, 1) })
-    def 'should throw an Exception if no reduce function provided'() {
-        when:
-        collection.group(~['x': 1],
-                         ~['x': 'b'],
-                         ~['z': 0],
-                         null);
-
-        then:
-        thrown(IllegalArgumentException)
-    }
 
     def 'should return null when findOne finds nothing'() {
         expect:
