+++ driver-legacy/src/test/unit/com/mongodb/DBCursorSpecification.groovy
@@ -27,10 +27,9 @@ import spock.lang.Specification
 
 import java.util.concurrent.TimeUnit
 
-import static com.mongodb.CustomMatchers.isTheSameAs
 import static Fixture.getMongoClient
+import static com.mongodb.CustomMatchers.isTheSameAs
 import static spock.util.matcher.HamcrestSupport.expect
-
 
 class DBCursorSpecification extends Specification {
 
@@ -124,8 +123,7 @@ class DBCursorSpecification extends Specification {
         then:
         expect executor.getReadOperation(), isTheSameAs(new FindOperation(collection.getNamespace(), collection.getObjectCodec())
                                                                 .filter(new BsonDocument())
-                                                                .projection(new BsonDocument())
-                                                                .modifiers(new BsonDocument()))
+                                                                .projection(new BsonDocument()))
     }
 
 
@@ -143,8 +141,7 @@ class DBCursorSpecification extends Specification {
         expect executor.getReadOperation(), isTheSameAs(new FindOperation(collection.getNamespace(), collection.getObjectCodec())
                                                                 .limit(-1)
                                                                 .filter(new BsonDocument())
-                                                                .projection(new BsonDocument())
-                                                                .modifiers(new BsonDocument()))
+                                                                .projection(new BsonDocument()))
     }
 
     def 'DBCursor methods should be used to create the expected operation'() {
@@ -190,7 +187,6 @@ class DBCursorSpecification extends Specification {
                 .partial(true)
                 .skip(1)
                 .sort(bsonSort)
-                .modifiers(new BsonDocument())
         )
 
         executor.getReadPreference() == readPreference
@@ -206,9 +202,7 @@ class DBCursorSpecification extends Specification {
         def filter = new BasicDBObject()
         def projection = BasicDBObject.parse('{a: 1, _id: 0}')
         def sort = BasicDBObject.parse('{a: 1}')
-        def modifiers = BasicDBObject.parse('{$comment: 1}')
         def bsonFilter = new BsonDocument()
-        def bsonModifiers = BsonDocument.parse(modifiers.toJson())
         def bsonProjection = BsonDocument.parse(projection.toJson())
         def bsonSort = BsonDocument.parse(sort.toJson())
         def comment = 'comment'
@@ -227,7 +221,6 @@ class DBCursorSpecification extends Specification {
                 .limit(1)
                 .maxAwaitTime(1, TimeUnit.MILLISECONDS)
                 .maxTime(1, TimeUnit.MILLISECONDS)
-                .modifiers(modifiers)
                 .noCursorTimeout(true)
                 .oplogReplay(true)
                 .partial(true)
@@ -257,7 +250,6 @@ class DBCursorSpecification extends Specification {
                 .limit(1)
                 .maxAwaitTime(1, TimeUnit.MILLISECONDS)
                 .maxTime(1, TimeUnit.MILLISECONDS)
-                .modifiers(bsonModifiers)
                 .noCursorTimeout(true)
                 .oplogReplay(true)
                 .partial(true)
@@ -273,45 +265,6 @@ class DBCursorSpecification extends Specification {
         )
 
         executor.getReadPreference() == findOptions.getReadPreference()
-    }
-
-    def 'DBCursor options should override DBCollectionFindOptions'() {
-        given:
-        def executor = new TestOperationExecutor([stubBatchCursor()]);
-        def collection = new DB(getMongoClient(), 'myDatabase', executor).getCollection('test')
-        def cursorType = CursorType.NonTailable
-        def readPreference = ReadPreference.primary()
-        def filter = new BasicDBObject()
-        def bsonFilter = new BsonDocument()
-        def findOptions = new DBCollectionFindOptions()
-                .cursorType(cursorType)
-                .noCursorTimeout(false)
-                .oplogReplay(false)
-                .partial(false)
-                .readPreference(readPreference)
-
-        def cursor = new DBCursor(collection, filter , findOptions)
-                .addOption(Bytes.QUERYOPTION_AWAITDATA)
-                .addOption(Bytes.QUERYOPTION_NOTIMEOUT)
-                .addOption(Bytes.QUERYOPTION_OPLOGREPLAY)
-                .addOption(Bytes.QUERYOPTION_PARTIAL)
-                .addOption(Bytes.QUERYOPTION_SLAVEOK)
-                .addOption(Bytes.QUERYOPTION_TAILABLE)
-
-        when:
-        cursor.toArray()
-
-        then:
-        expect executor.getReadOperation(), isTheSameAs(new FindOperation(collection.getNamespace(), collection.getObjectCodec())
-                .modifiers(new BsonDocument())
-                .cursorType(CursorType.TailableAwait)
-                .filter(bsonFilter)
-                .noCursorTimeout(true)
-                .oplogReplay(true)
-                .partial(true)
-        )
-
-        executor.getReadPreference() == ReadPreference.secondaryPreferred()
     }
 
     def 'count should create the correct CountOperation'() {
